fix: cast bufsize and byte_left to int for SSL_read/SSL_write calls
SSL_read and SSL_write take int for the length argument; passing size_t
triggers -Wconversion. Add explicit (int) casts, consistent with the
existing (int)bufsize cast already present on the Windows recv path.

+++ upnp/src/genlib/net/sock.c
@@ -241,7 +241,7 @@ static int sock_read_write(SOCKINFO *info,
 #ifdef UPNP_ENABLE_OPEN_SSL
 			if (info->ssl) {
 				numBytes = (long)SSL_read(
-					info->ssl, buffer, bufsize);
+					info->ssl, buffer, (int)bufsize);
 			} else {
 #endif
 				/* read data. */
@@ -263,7 +263,7 @@ static int sock_read_write(SOCKINFO *info,
 				if (info->ssl) {
 					num_written = SSL_write(info->ssl,
 						buffer + bytes_sent,
-						byte_left);
+						(int)byte_left);
 				} else {
 #endif
 					/* write data. */
